+++ services/web/server/tests/conftest.py
@@ -22,11 +22,6 @@ log = logging.getLogger(__name__)
 logging.getLogger("openapi_spec_validator").setLevel(logging.WARNING)
 logging.getLogger("sqlalchemy").setLevel(logging.WARNING)
 
-## HELPERS
-sys.path.append(str(current_dir / "helpers"))
-
-
-## FIXTURES: standard paths
 
 
 @pytest.fixture(scope="session")
@@ -35,30 +30,6 @@ def package_dir() -> Path:
     dirpath = Path(simcore_service_webserver.__file__).resolve().parent
     assert dirpath.exists()
     return dirpath
-
-
-@pytest.fixture(scope="session")
-def osparc_simcore_root_dir() -> Path:
-    """ osparc-simcore repo root dir """
-    WILDCARD = "services/web/server"
-
-    root_dir = Path(current_dir)
-    while not any(root_dir.glob(WILDCARD)) and root_dir != Path("/"):
-        root_dir = root_dir.parent
-
-    msg = f"'{root_dir}' does not look like the git root directory of osparc-simcore"
-    assert root_dir.exists(), msg
-    assert any(root_dir.glob(WILDCARD)), msg
-    assert any(root_dir.glob(".git")), msg
-
-    return root_dir
-
-
-@pytest.fixture(scope="session")
-def env_devel_file(osparc_simcore_root_dir) -> Path:
-    env_devel_fpath = osparc_simcore_root_dir / ".env-devel"
-    assert env_devel_fpath.exists()
-    return env_devel_fpath
 
 
 @pytest.fixture(scope="session")
